+++ docs/CHANGELOG.md
@@ -1,6 +1,20 @@
 CHANGELOG
 ---------
 
+#### Next release
+- Update dependencies:
+  - Gradle to 7.4.2
+  - Android Gradle plugin to 7.2.1
+  - Build tool to 33.0.0
+  - Compile and Target SDK to 32
+  - Kotlin to 1.7.0, Coroutines to 1.6.2, Kover to 0.5.1
+  - AndroidX AppCompat to 1.4.2, Core to 1.8.0, Constraint Layout to 2.1.4
+  - Firebase BOM to 30.1.4, Crashlytics Gradle to 2.9.0
+  - Hilt to 2.42
+  - Material Component to 1.6.1
+  - Mockk to 1.12.4
+  - Roblectric to 4.8.1
+
 #### v0.21.0
 - New features:
   - Support cross references (#21)